fix: chadchartWalkout now always appears on screen
ChadchartWalkout (the node in old commits) "doesn't appear" because when it spawns, it immediately collides with player, so it calls `super.collect()` and will hide itself. i fixed this by creating a new `chadchart_walkout.tscn` which is not `Collectable`. and use that instead.

also properly create & delete chadchartWalkout node now